Data persistence implementation
Implemented Hiscore list with 5 best scores, start button, exit button and current player name field shown on the start menu and esc key handler to return to the start menu from main scene.
Data persistence between sessions: hiscore list is saved in OnApplicationQuit() and loaded on menu scene Start() (or when HiscoreManager object is created)
Data persistence between scenes : highest score is shown when playing the game and if the player breaks the score, current player name is added to the list.